Changed the console buffer to auto size depending on the window and font sizes. It now handles all window sizes properly

+++ Nildavellir/Graphics.h
@@ -29,7 +29,7 @@ public:
 	void Shutdown();
 
 	//
-	void Render(double timeElapsed);
+	void Render(float timeElapsed);
 
 	//
 	void SetWindowSize( unsigned short width, unsigned short height);
